Update README.md
The program given below creates a simple calculator in C++ programming that performs four basic mathematical operations, such as addition, subtraction, multiplication, and division, depending on the user's choice.

To do the task of the calculator, first receive the choice and then any two numbers (if the choice is not to exit). Then, as shown in the program below, use the switch case to identify the choice and perform the necessary actions

A calculator program that performs basic arithmetic operations such as addition, subtraction, multiplication, and division. Allow the user to input two numbers and choose an operation to perform.